requirements.txt
fastapi
uvicorn[standard]
easyocr
opencv-python-headless
numpy
python-dotenv
openai
pymongo
